Made valid email method
#20
