+++ cogs/errors.py
@@ -4,6 +4,8 @@ import discord
 from discord.ext import commands
 from discord.ext.commands import errors as cmderr
 
+from util.email import is_valid_email
+
 
 class Errors(commands.Cog):
 	def __init__(self, bot):
@@ -34,24 +36,6 @@ class Errors(commands.Cog):
 				await self.bot.invoke(new_ctx)
 		else:
 			print(exception)
-
-
-def is_valid_email(email: str):
-	try:
-		dm = email.split('@')[1]  # split the string based on the @ symbol
-	except AttributeError:
-		return False
-
-	if not dm:
-		return False
-
-	if len(email.split('@')[0]) > 64 or len(email.split('@')[1]) > 255:  # valid emails have 64char max before @, 255 max after
-		return False
-
-	if set('+').intersection(email):  # to prevent people from making extra email addresses
-		return False
-
-	return True
 
 
 def setup(bot):
